Check QDT's local time abbreviation as well as the qTzName()s
Since QDateTime uses some fall-backs if qTzName() doesn't give it
something useful (as happens on MS-Win when local time is UTC),
QDateTimeParser should check the result of those fall-backs as well as
the qTzName()s when checking for local-time as zone.

In 6.5, startsWithLocalTimeZone() was a class member of QDTP
implemented in qlocaltime.cpp, to share qt_tzname(), which hadn't yet
become qTzName(). So adapt the changes to apply to it instead.

Change-Id: Ic809b7e44cd0c83fb076b24c27547268345fa379
Reviewed-by: Konrad Kujawa <konrad.kujawa@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit c888e3922d73df791f0f31553536abf03b241a65)