tst_QSortFilterProxyModel: fix mem-leaks II: sortStable()
Like in 9f8449a054a165ae560c91d3e31409bc293df80d, for
doubleProxySelectionSetSourceModel(), the sortStable() test also
leaked _everything_. Fix in the same way, by allocating model and view
on the stack intead of the heap.

With this patch, tst_QSortFilterProxyModel is now asan-, but not
ubsan-clean (and, because of
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=110704, asan still
reports a leak in cp-demangle.c, which should be™ gone once we fix the
rest of QTBUG-99563).

Task-number: QTBUG-115264
Change-Id: Ic0e833d7336435e324457f9d9667ee8573a7dafc
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 65647d54b9fddb0b0e3238024d79e12e4bf53295)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>